Upgraded UnitTest++ to version 1.4

+++ UnitTest++/src/Posix/TimeHelpers.cpp
@@ -14,18 +14,18 @@ void Timer::Start()
     gettimeofday(&m_startTime, 0);
 }
 
-
-int Timer::GetTimeInMs() const
+double Timer::GetTimeInMs() const
 {
     struct timeval currentTime;
     gettimeofday(&currentTime, 0);
-    int const dsecs = currentTime.tv_sec - m_startTime.tv_sec;
-    int const dus = currentTime.tv_usec - m_startTime.tv_usec;
-    return dsecs*1000 + dus/1000;
+
+	double const dsecs = currentTime.tv_sec - m_startTime.tv_sec;
+    double const dus = currentTime.tv_usec - m_startTime.tv_usec;
+
+	return (dsecs * 1000.0) + (dus / 1000.0);
 }
 
-
-void TimeHelpers::SleepMs (int ms)
+void TimeHelpers::SleepMs(int ms)
 {
     usleep(ms * 1000);
 }
